changed styles and buttons
