+++ client/src/App.js
@@ -3,17 +3,21 @@ import { BrowserRouter, Routes, Route, Link } from 'react-router-dom'
 import Home from './pages/HomePage/HomePage';
 import UploadAnimal from './pages//UploadAnimalPage/UploadAnimal';
 import styles from './styles/App.module.css';
+import NavButton from './components/common/NavButton/NavButton';
 
 function App() {
   return (
     <BrowserRouter>
       <div className={styles.app}>
         <header className={styles.appHeader}>
-          <img src="/logo2.png" alt="Logo" className={styles.appLogo} />
+          <img src="/logo.png" alt="Logo" className={styles.appLogo} />
           <nav className={styles.appNav}>
             <Link to="/" className={styles.appLink}>
-              <button class="log">Login</button>
-              <button class="reg">Sign up</button>
+              <NavButton>Login</NavButton>
+              <NavButton>Sign up</NavButton>
+            </Link>
+            <Link to="/UploadAnimal" className={styles.appLink}>
+              <NavButton>Upload Animal</NavButton>
             </Link>
           </nav>
         </header>
