+++ client/src/components/common/NavButton/NavButton.module.css
@@ -1,14 +1,17 @@
 .button {
   padding: 10px 20px;
-  background-color: #FF7D9C;
-  color: white;
+  background-color: #664343;
+  color: #FFF0D1;
   border: none;
   border-radius: 4px;
   cursor: pointer;
   font-size: 16px;
   width: auto;
+  transition: background-color 0.3s ease outline-color 0.3s ease;
+  outline: 1px solid #FFF0D1;
 }
 
 .button:hover {
-  background-color: #ffffff;
+  background-color: #FFF0D1;
+  color: #795757;
 }
