+++ client/src/context/AnimalContext.js
@@ -26,4 +26,22 @@ const AnimalProvider = ({ children }) => {
     );
 };
 
-export { AnimalContext, AnimalProvider };
+const animalOptions = [
+    { value: "", label: "", disabled: true, hidden: true }, 
+    { value: "Dog", label: "Dog" },
+    { value: "Cat", label: "Cat" },
+    { value: "Rabbit", label: "Rabbit" },
+    { value: "Guinea Pig", label: "Guinea Pig" },
+    { value: "Hamster", label: "Hamster" },
+    { value: "Bird", label: "Bird" },
+    { value: "Fish", label: "Fish" },
+    { value: "Reptile", label: "Reptile" },
+    { value: "Horse", label: "Horse" },
+    { value: "Goat", label: "Goat" },
+    { value: "Chicken", label: "Chicken" },
+    { value: "Other", label: "Other" },
+  ];
+  
+ 
+
+export { AnimalContext, AnimalProvider, animalOptions };
