+++ client/src/pages/UploadAnimalPage/UploadAnimal.module.css
@@ -1,14 +1,3 @@
-.upload {
-  text-align: center;
-  padding: 20px;
-}
-
-.headline {
-  font-size: 2em;
-  margin-bottom: 20px;
-  color: #53474F;
-}
-
 .image { 
   max-height: 400px;
   max-width: 100%;
