+++ client/src/styles/global.css
@@ -19,7 +19,7 @@ body {
   --error: #e7195a;
 }
 body {
-  background: #f1f1f1;
+  background: #FFF0D1;
   margin: 0;
   font-family: "Poppins";
 }
